pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "invoice-ai"
version = "0.1.0"
description = "AI-powered invoice and receipt data extraction using Claude's vision capabilities"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "invoice-ai contributors" },
]
keywords = ["invoice", "receipt", "ocr", "ai", "claude", "extraction", "pdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "anthropic>=0.39.0",
    "click>=8.1.0",
    "pdfplumber>=0.11.0",
    "pdf2image>=1.17.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
invoice-ai = "invoice_ai.cli:cli"

[project.urls]
Homepage = "https://github.com/dom-ran/invoice-ai"
Repository = "https://github.com/dom-ran/invoice-ai"
Issues = "https://github.com/dom-ran/invoice-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/invoice_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
